+++ R/redcap.R
@@ -743,7 +743,7 @@ ody_rc_import <- function(
     )
   }
 
-  cli::cli_alert_info("Importing data from RedCap")
+  cli::cli_alert_info("Importing data from REDCap")
   rc_raw_import <- import_rc(token, form, url)
 
   if (!label && !nest && is.null(form)) return(rc_raw_import)
@@ -1554,3 +1554,51 @@ ody_rc_add_site <- function(tbl,
     dplyr::relocate("site", .before = position)
 
 }
+
+
+#' Adds labels to a data frame using REDCap metadata.
+#'
+#' This function takes a data frame and applies variable labels derived from
+#' the REDCAp metadata column `field_label`. If no `redcap_data` from
+#' which to extract the metadata is provided, it will look for an object named
+#' `redcap_data` in the global environment.
+#'
+#' @param df A data frame to which the labels will be added.
+#' @param redcap_data Optional. A REDCap data frame containing the metadata as
+#' attribute. If not supplied, the function will use `redcap_data` from the
+#' global environment.
+#'
+#' @return A data frame with variable labels applied from the REDCap metadata.
+#' @export
+ody_rc_add_label <- function(df, redcap_data = NULL) {
+
+  if (is.null(redcap_data)) {
+    if (!exists("redcap_data", envir = .GlobalEnv)) {
+      stop("No redcap_data object found.")
+    }
+    redcap_data <- get("redcap_data", envir = .GlobalEnv)
+  }
+
+  var_names <- names(df)
+  labels <-
+    attr(redcap_data, "metadata") |>
+    dplyr::filter(
+      .data[["field_name"]] %in% var_names,
+      !is.na(.data[["field_label"]])
+    ) |>
+    dplyr::select("field_name", "field_label")
+
+  lab_lang <- stringr::str_c(
+    "list(",
+    stringr::str_c(
+      labels$field_name, " = '", labels$field_label, "'", collapse = ", "
+    ),
+    ")") |>
+    str2lang()
+
+  labelled::var_label(df) <- eval(lab_lang)
+
+  df
+
+}
+